requirements.txt
streamlit
openai-whisper
nltk
spacy
transformers
torch
fpdf
ffmpeg-python
